move the kubeconfig operator to kube lib
